created custom hook to simulate progress
