+++ src/components/Progress/Progress.stories.tsx
@@ -1,5 +1,6 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import Progress from './Progress';
+import useSimulateProgress from '../../utils/hooks/useSimulateProgress';
 
 export default {
   title: 'Progress',
@@ -7,70 +8,24 @@ export default {
 };
 
 export const DefaultProgress = () => {
-  // linear progress bar width
-  const [width, setWidth] = useState<number>(0);
+  const progress = useSimulateProgress();
 
-  // set interval to simulate progress
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setWidth((currWidth) => {
-        // reset after reaching 100
-        if (currWidth === 100) {
-          return 0;
-        }
-
-        // create random value
-        const diff = Math.random() * 10;
-
-        // return whichever is smaller
-        return Math.min(currWidth + diff, 100);
-      });
-    }, 400);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, []);
-
-  return <Progress pgValue={width} />;
+  return <Progress pgValue={progress} />;
 };
 
 export const DiffTypeProgress = () => {
-  // linear progress bar width
-  const [width, setWidth] = useState<number>(0);
-
-  // set interval to simulate progress
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setWidth((currWidth) => {
-        // reset after reaching 100
-        if (currWidth === 100) {
-          return 0;
-        }
-
-        // create random value
-        const diff = Math.random() * 10;
-
-        // return whichever is smaller
-        return Math.min(currWidth + diff, 100);
-      });
-    }, 400);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, []);
+  const progress = useSimulateProgress();
 
   return (
     <>
       <h1>Linear</h1>
       <br />
-      <Progress pgValue={width} />
+      <Progress pgValue={progress} />
       <br />
 
       <h1>Circular</h1>
       <br />
-      <Progress pgType="circular" pgValue={width} />
+      <Progress pgType="circular" pgValue={progress} />
       <br />
     </>
   );
